# frontend/src/api/openapi.merged.yaml
openapi: 3.0.0
info:
  title: ApeRAG API
  description: ApeRAG API Documentation
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /invite:
    post:
      summary: Create an invitation
      description: Create a new invitation for a user to register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/invitationCreate'
      responses:
        '200':
          description: Invitation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/invitation'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /invitations:
    get:
      summary: Get all invitations
      description: Get all invitations
      responses:
        '200':
          description: Invitations fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/invitationList'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /register:
    post:
      summary: Register a new user
      description: Register a new user with an invitation token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/register'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /login:
    post:
      summary: Login a user
      description: Login a user with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/login'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
...
  /users:
    get:
      summary: Get all users
      description: Get all users
      responses:
        '200':
          description: Users fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userList'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /user:
    get:
      summary: Get user info
      description: Get user info
      responses:
        '200':
          description: User info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /change-password:
    post:
      summary: Change user password
      description: Change user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/changePassword'
      responses:
        '200':
          description: Password changed successfully
  /users/{user_id}:
    delete:
      summary: Delete a user
      description: Delete a user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted successfully
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
    CookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
  schemas:
...
    pageResult:
      description: PageResult info
      properties:
        page_number:
          type: integer
          description: The page number
        page_size:
          type: integer
          description: The page size
        count:
          type: integer
          description: The total count of items
      type: object
...
    failResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: '400'
        message:
          type: string
          description: Error message
          example: Invalid request
...
    invitationCreate:
      type: object
      properties:
        username:
          type: string
          description: The username of the user
        email:
          type: string
          description: The email of the user
        role:
          type: string
          description: The role of the user (admin, rw, ro)
          enum:
            - admin
            - rw
            - ro
    invitation:
      type: object
      properties:
        email:
          type: string
          description: The email of the user
        token:
          type: string
          description: The token of the invitation
        created_by:
          type: string
          description: The ID of the user who created the invitation
        created_at:
          type: string
          description: The date and time the invitation was created
        is_valid:
          type: boolean
          description: Whether the invitation is valid
        used_at:
          type: string
          description: The date and time the invitation was used
        role:
          type: string
          description: The role of the user (admin, rw, ro)
          enum:
            - admin
            - rw
            - ro
        expires_at:
          type: string
          description: The date and time the invitation will expire
    invitationList:
      type: object
      description: A list of invitations
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/invitation'
        pageResult:
          $ref: '#/components/schemas/pageResult'
    register:
      type: object
      description: The email of the user
      properties:
        token:
          type: string
          description: The invitation token
        email:
          type: string
          description: The email of the user
        username:
          type: string
          description: The username of the user
        password:
          type: string
          description: The password of the user
    user:
      type: object
      properties:
        id:
          type: string
          description: The ID of the user
        username:
          type: string
          description: The username of the user
        email:
          type: string
          description: The email of the user
        role:
          type: string
          description: The role of the user
        is_active:
          type: boolean
          description: Whether the user is active
        date_joined:
          type: string
          description: The date and time the user joined the system
    login:
      type: object
      properties:
        username:
          type: string
          description: The username of the user
        password:
          type: string
          description: The password of the user
    userList:
      type: object
      description: A list of users
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/user'
        pageResult:
          $ref: '#/components/schemas/pageResult'
    changePassword:
      type: object
      properties:
        username:
          type: string
          description: The username of the user
        old_password:
          type: string
          description: The old password of the user
        new_password:
          type: string
          description: The new password of the user